+++ src/fintrist/models.py
@@ -361,7 +361,7 @@ class Study(BaseStudy):
         self.data, newalerts = function(**parent_data, **self.params)
         if self.alert_overwrite(arrow.get(self.timestamp, Config.TZ)):
             self.alertslog.remove_alert()
-        self.timestamp = arrow.get(self.timestamp, Config.TZ).datetime
+        self.timestamp = arrow.now(Config.TZ).datetime
         self.alertslog.record_alerts(newalerts, self.timestamp)
         self.save()