+++ src/fintrist/services.py
@@ -126,3 +126,11 @@ def spawn_stream(stream_name, **kwargs):
         spawn_study(recipe, **kwargs) for recipe in stream_obj.recipes
     ]
     return newstudies
+
+def create_backtest(name, period='1y'):
+    """Create a Study that can be used to backtest."""
+    model = get_study(name)
+    backtest = create_study(
+        f"{name} backtest", 'backtest',
+        parents={'model': model}, params={'period': period})
+    return backtest